board/stmicroelectronics/stm32mp1/post-build.sh: add symlinks to WiFi firmware
Recent versions of the Broadcom WiFi driver first tries to load the
WiFi firmware using a firmware file name suffixed by the compatible
string of the platform, before falling back to the plain firmware
name. In order to avoid strange looking firmware loading errors during
the boot process, we create the appropriate symbolic links.

As part of this, we slightly refactor the script into two
functions. bash is now used instead of sh to be able to use
pushd/popd.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>