docs/qt5.md: Add qt5 packages and instructions for stm32mp257
This demo is almost complete. The only missing piece is the GPU driver provided
by the gcnano-binaries package. Currently, the stm32mp-oem version of this
package has support for stm32mp257 platforms, but this is not the case for the
upstream package which is integrated in Buildroot.

Add QT5 demo instructions and packages.

Signed-off-by: Romain Gantois <romain.gantois@bootlin.com>