board/stm32mp2/genimage: Restore metadata partitions and UUIDs and add precompiled metadata
This reverts commit a3600ff0a2e8a106bf0309d45ea016b97f391e05.

We've planned to support FWU in this external eventually so keep this
partition layout for now so we don't have to reconfigure TF-A U-Boot and
Rauc later on.

The firmware update metadata used by TF-A is no longer generated by the python
tool integrated in the TF-A sources. They are now generated by the mkfwumdata
tool in the U-Boot sources. Prebuild a metadata binary using mkfwumdata and use
it as a default until we properly support the FWU feature in the demo
configuration.