board/stmicroelectronics/stm32mp1: create proper U-Boot env partitions
These partitions match the U-Boot configuration in terms of where the
environment is stored.

The extlinux.conf are changed to use PARTLABEL= instead to not depend
on the partition numbering. For /etc/fstab, support of labels is not
enabled in Busybox, so we have to continue to use a partition number,
which gets updated.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>